# openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BhashaAI Backend",
    "description": "\n## BhashaAI - Generative AI for Education & Gujarati Language Technologies\n\nA comprehensive API for educational AI services supporting:\n\n- **Question Paper Generation** - AI-powered question paper creation from educational content\n- **Assignment Help** - Socratic tutoring and step-by-step solutions\n- **Teaching Tools** - Mind maps, slides, and study materials generation\n- **Paper Checking** - OCR-based automated evaluation with Gujarati support\n- **Gujarati Learning** - Vocabulary, grammar, and gamified learning\n- **Audio Services** - STT/TTS for Gujarati language\n\n### Authentication\nAll protected endpoints require JWT Bearer token authentication.\n\n### Bilingual Support\nResponses include both English and Gujarati (ગુજરાતી) messages.\n        ",
    "contact": {
      "name": "BhashaAI Team",
      "email": "support@bhashaai.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/assignments/submit": {
      "post": {
        "tags": [
          "Assignments",
          "Assignments"
        ],
        "summary": "Submit Assignment",
        "description": "Submit a question for solution or help.",
        "operationId": "submit_assignment_api_v1_assignments_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/assignments": {
      "post": {
        "tags": [
          "Assignments",
          "Assignments"
        ],
        "summary": "Create Assignment",
        "description": "Publish a question paper as an assignment.",
        "operationId": "create_assignment_api_v1_assignments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Assignments",
          "Assignments"
        ],
        "summary": "List Assignments",
        "description": "List user's assignments.",
        "operationId": "list_assignments_api_v1_assignments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assignments/{assignment_id}": {
      "get": {
        "tags": [
          "Assignments",
          "Assignments"
        ],
        "summary": "Get Assignment Details",
        "description": "Get assignment details including solution/help session.",
        "operationId": "get_assignment_api_v1_assignments__assignment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assignments/{assignment_id}/hint": {
      "post": {
        "tags": [
          "Assignments",
          "Assignments"
        ],
        "summary": "Get Next Hint",
        "description": "Get next hint for a help session.\nOnly valid if assignment mode is 'help'.",
        "operationId": "get_hint_api_v1_assignments__assignment_id__hint_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HintRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response data payload"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Human-readable message (English)"
          },
          "message_gu": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Gu",
            "description": "Human-readable message (Gujarati / ગુજરાતી)"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "APIResponse",
        "description": "Standard API response envelope.\n\nAll API responses follow this structure for consistency.\n\nAttributes:\n    success: Whether the operation was successful\n    data: Response payload (type varies by endpoint)\n    message: Human-readable message in English\n    message_gu: Human-readable message in Gujarati\n\nExample:\n    ```json\n    {\n        \"success\": true,\n        \"data\": {\"id\": \"uuid-here\", \"name\": \"Example\"},\n        \"message\": \"Operation successful\",\n        \"message_gu\": \"કામગીરી સફળ\"\n    }\n    ```"
      },
...
      "AssignmentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "question_paper_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Paper Id"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "published"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "solve"
          }
        },
        "type": "object",
        "required": [
          "title",
          "question_paper_id"
        ],
        "title": "AssignmentCreate",
        "description": "Schema for creating a new assignment from a QP."
      },
      "AssignmentMode": {
        "type": "string",
        "enum": [
          "solve",
          "help"
        ],
        "title": "AssignmentMode",
        "description": "Assignment solving mode."
      },
      "AssignmentSubmit": {
        "properties": {
          "input_type": {
            "$ref": "#/components/schemas/InputType",
            "default": "text"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "grade_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Level"
          },
          "mode": {
            "$ref": "#/components/schemas/AssignmentMode",
            "default": "solve"
          },
          "language": {
            "type": "string",
            "pattern": "^(gu|en|gu-en)$",
            "title": "Language",
            "default": "gu"
          },
          "question_text": {
            "type": "string",
            "minLength": 1,
            "title": "Question Text"
          },
          "question_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Image Url"
          }
        },
        "type": "object",
        "required": [
          "question_text"
        ],
        "title": "AssignmentSubmit",
        "description": "Schema for submitting an assignment."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HintRequest": {
        "properties": {
          "student_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Response"
          },
          "request_next_level": {
            "type": "boolean",
            "title": "Request Next Level",
            "default": true
          }
        },
        "type": "object",
        "title": "HintRequest",
        "description": "Request for next hint in help mode."
      },
      "InputType": {
        "type": "string",
        "enum": [
          "text",
          "image",
          "pdf"
        ],
        "title": "InputType",
        "description": "Assignment input types."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /api/v1/auth/login"
      },
      "RefreshToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Refresh-Token",
        "description": "JWT refresh token for obtaining new access tokens"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.bhashaai.com",
      "description": "Production server"
    }
  ]
}